use int16_t calc for bounce, velocity should be only higher than 128 if its to be used as the offset for linked actors.

+++ DynamicActorPluginExample/plugins/DynamicActorPlugin/engineAlt/Cont_Scr_M/src/dynamic_actor.c
@@ -487,7 +487,7 @@ void dynamic_actor_update(void) BANKED {
                     if (def->bounce == 255) {
                         actor->actor_vel_y = -actor->actor_vel_y;
                     } else {
-                        actor->actor_vel_y = -(WORD)(((int32_t)actor->actor_vel_y * def->bounce) >> 8);
+                        actor->actor_vel_y = -(WORD)(((int16_t)actor->actor_vel_y * def->bounce) >> 8);
                     }
                     // Kill micro-bounces caused by gravity pumping while resting
                     if (moving_down && (-actor->actor_vel_y <= def->gravity)) {
